Respond to review comments.
Move SyncToAsyncCMgr.  Use firethread.
