+++ synapse/tests/test_axon.py
@@ -13,6 +13,8 @@ import synapse.lib.threads as s_threads
 
 import synapse.tests.common as s_test
 
+from synapse.tests.utils import SyncToAsyncCMgr
+
 logger = logging.getLogger(__name__)
 
 # This causes blocks which are not homogeneous when sliced in kibibyte lengths
@@ -38,7 +40,7 @@ class AxonTest(s_test.SynTest):
     async def test_blobstor(self):
         with self.getTestDir() as dirn:
             path0 = os.path.join(dirn, 'blob0')
-            async with s_test.SyncToAsyncCMgr(s_axon.BlobStor, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as bst0:
+            async with SyncToAsyncCMgr(s_axon.BlobStor, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as bst0:
 
                 data1 = b'asdfqwerhehehaha'
 
@@ -71,7 +73,7 @@ class AxonTest(s_test.SynTest):
 
                 path1 = os.path.join(dirn, 'blob1')
                 blobstorconf = {'mapsize': s_test.TEST_MAP_SIZE}
-                async with s_test.SyncToAsyncCMgr(s_axon.BlobStor, path1, conf=blobstorconf) as bst1:
+                async with SyncToAsyncCMgr(s_axon.BlobStor, path1, conf=blobstorconf) as bst1:
 
                     clone_data = [x async for x in bst0.clone(0)]
                     await bst1._consume_clone_data(clone_data)
@@ -93,7 +95,7 @@ class AxonTest(s_test.SynTest):
     async def test_blobstor_stat(self):
         with self.getTestDir() as dirn:
             path0 = os.path.join(dirn, 'blob0')
-            async with s_test.SyncToAsyncCMgr(s_axon.BlobStor, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as bst0:
+            async with SyncToAsyncCMgr(s_axon.BlobStor, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as bst0:
 
                 blobs = (
                     (None, 0, os.urandom(1000)),
@@ -117,7 +119,7 @@ class AxonTest(s_test.SynTest):
     async def test_blobstor_metrics(self):
         with self.getTestDir() as dirn:
             path0 = os.path.join(dirn, 'blob0')
-            async with s_test.SyncToAsyncCMgr(s_axon.BlobStor, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as bst0:
+            async with SyncToAsyncCMgr(s_axon.BlobStor, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as bst0:
 
                 blobs = (
                     (None, 0, os.urandom(1000)),
@@ -154,7 +156,7 @@ class AxonTest(s_test.SynTest):
     @s_glob.synchelp
     async def test_blobstor_remote(self):
         with self.getTestDir():
-            async with s_test.SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
+            async with SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
                     await dmon._getTestProxy('blobstor00') as bst0:
                 stats = await bst0.stat()
                 self.eq(stats, {})
@@ -166,9 +168,9 @@ class AxonTest(s_test.SynTest):
     async def test_axon(self):
         with self.getTestDir() as dirn:
             path0 = os.path.join(dirn, 'axon0')
-            async with s_test.SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
+            async with SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
                     await dmon._getTestProxy('blobstor00') as blobstor0, \
-                    s_test.SyncToAsyncCMgr(s_axon.Axon, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as axon:
+                    SyncToAsyncCMgr(s_axon.Axon, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as axon:
 
                 self.eq((), [x async for x in axon.metrics()])
                 self.eq((), [x async for x in await blobstor0.metrics()])
@@ -213,7 +215,7 @@ class AxonTest(s_test.SynTest):
     @s_glob.synchelp
     async def test_axon_remote(self):
         with self.getTestDir():
-            async with s_test.SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
+            async with SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
                     await dmon._getTestProxy('axon00') as axon:
                 blobstorurl = f'tcp://{dmon.addr[0]}:{dmon.addr[1]}/blobstor00'
                 await axon.addBlobStor(blobstorurl)
@@ -230,8 +232,8 @@ class AxonTest(s_test.SynTest):
     async def test_axon_uploader(self):
         with self.getTestDir() as dirn:
             path0 = os.path.join(dirn, 'axon0')
-            async with s_test.SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
-                    s_test.SyncToAsyncCMgr(s_axon.Axon, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as axon:
+            async with SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
+                    SyncToAsyncCMgr(s_axon.Axon, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as axon:
                 abhash = hashlib.sha256(b'ab').digest()
                 cdhash = hashlib.sha256(b'cd').digest()
 
@@ -270,9 +272,9 @@ class AxonTest(s_test.SynTest):
         ''' Bring up an axon and a blobstor, then bring up a second blobstor that's cloning the first '''
         with self.getTestDir() as dirn:
             path0 = os.path.join(dirn, 'axon0')
-            async with s_test.SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
+            async with SyncToAsyncCMgr(self.getTestDmon, mirror='axondmon') as dmon, \
                     await dmon._getTestProxy('blobstor00') as blobstor0, \
-                    s_test.SyncToAsyncCMgr(s_axon.Axon, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as axon:
+                    SyncToAsyncCMgr(s_axon.Axon, path0, conf={'mapsize': s_test.TEST_MAP_SIZE}) as axon:
 
                 blobstorurl0 = f'tcp://{dmon.addr[0]}:{dmon.addr[1]}/blobstor00'
                 blobstorurl1 = f'tcp://{dmon.addr[0]}:{dmon.addr[1]}/blobstor01'
@@ -282,7 +284,7 @@ class AxonTest(s_test.SynTest):
                 blobstor1conf = {'mapsize': s_test.TEST_MAP_SIZE, 'cloneof': blobstorurl0}
 
                 # Make a second blobstor that clones the first
-                async with s_test.SyncToAsyncCMgr(s_axon.BlobStor, path0, conf=blobstor1conf) as blobstor1:
+                async with SyncToAsyncCMgr(s_axon.BlobStor, path0, conf=blobstor1conf) as blobstor1:
                     # Make sure the cloning works
                     for i in range(10):
                         if blobstor1.getCloneProgress() >= 2:
